requirements.txt
opencv-python==4.5.1.48
opencv-contrib-python==4.4.0.46
numpy==1.19.3
Pillow==8.1.0
imutils==0.5.3
tensorflow==2.4.0